requirements.txt
pandas==2.2.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==0.21.1
selenium==4.39.0
openpyxl==3.1.2
